pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "krknctl-ai"
version = "0.1.0"
description = "Chaos engineering command generation using Qwen3-1.7B + LoRA"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.51.0",
    "peft>=0.7.0",
    "accelerate>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
krknctl-ai = "krknctl_ai.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
